Support native GnuTLS for STARTTLS
Emacs 24 supports linking to the GnuTLS library.  Let's use it when
available.

Also add a customisable variable for ignoring invalid certificates.
We should now be validating certificates against the XMPP server name,
not the hostname from DNS SRV, so there should be less need for this
now, but there's always the occasional basement server with a
self-signed certificate...