Remove most autoload cookies for defcustoms
They shouldn't be used just to have them listed in Customize; see
http://debbugs.gnu.org/cgi/bugreport.cgi?bug=14247 .  Besides, they
should be loaded by the requires in jabber.el anyway.

I left jabber-account-list and jabber-display-menu in for now, as they
are involved in a complicated dance regarding whether to display the
Jabber menu by default.  Need to solve this somehow and get rid of
those two autoload cookies as well.