Use original JID when choosing which accounts to reconnect
Some Jabber servers (notably Facebook Chat) assign a different JID at
resource binding.  This used to confuse jabber-connect-all, such that it
would establish a new connection although the existing one is perfectly
healthy.

With this change, we save the JID that we originally connected with, and
use that to avoid duplicate connections.