Chat prompts are fields; use for goto-address
Give all chat prompts a 'field property.  This changes behaviour,
hopefully to the better.  For example, C-a will now stop at the
beginning of the chat text; use C-b C-a at that point to go to the
beginning of the prompt.

Use the field property to improve jabber-chat-goto-address.  Previously,
we would try to make URLs clickable in the entire buffer, which besides
being wasteful would cease working once the buffer size was past
goto-address-fontify-maximum-size.