Don't display menu bar indicator when installed
See "(elisp) Coding Conventions".  Displaying a menu bar indicator if
the package was installed violates its first convention, asides from
that this check yields a truthy value in nearly all cases for people
using this package and therefore makes the other possible branch (is
this mode actually in use?) unused.