Don't signal error for chat states on dead connection
If there is a chat buffer with chat state notifications enabled, and the
connection being used has been lost, and the user has reconnected, and
started typing a message in the chat buffer, then we would try to send
the "typing" notification to the old dead connection, causing an error.
Avoid that by ignoring errors when sending notifications.