Extract jabber-xml-parse-next-stanza, and test it
Move the functionality of reading the next complete stanza into a
separate function, and add some test cases for it.  I'd like to get rid
of jabber-xml-skip-tag-forward at some point, so a good first step
towards that should be moving calls to it into a function that's easy to
test in isolation.