Don’t base64-decode when there’s nothing to decode
Servers are free to return the string “=” instead of actual base64 data in the
success message payload as per [1] (cf. [2]). Prosody does this as of v0.10.
Hence, despite login success, we choke when trying to base64-decode non-base64
data and get stuck after auth.

We avoid getting stuck by making sure the data is not just nil but also not a
void “=”.

[1] <http://xmpp.org/rfcs/rfc6120.html#sasl-process-neg-initiate>
[2] <https://prosody.im/issues/issue/729>

Fixes #19